Add files via upload
Write algorithms and programs to simulate a deck of cards and deal hands of cards. Conventional deck of cards has 52 cards: 13 ranks and 4 suits. Your simulation should be able to create a deck of cards, display the original deck, shuffle the deck, display the shuffled deck, deal (from the top of the deck) the specified number of cards to the specified number of players (as per the command-line) and display each of the hands of cards. Execution would look something like: ./MainCardHands.out 7 5